% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata.R
\docType{data}
\name{exampledata}
\alias{exampledata}
\title{A subset of the 1992 National Adult Literacy Study (NALS) prison study
  public-use microdata file.}
\format{
An object of class \code{"data.frame"};
}
\usage{
data(exampledata)
}
\description{
A subset of the 1992 National Adult Literacy Study (NALS) prison study
  public-use microdata file. It has 20 variables and 182 records.
}
\keyword{datasets}
