% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loglinear.R
\name{sdc_loglinear_iter}
\alias{sdc_loglinear_iter}
\alias{print.sdc_loglinear_iter}
\alias{plot.sdc_loglinear_iter}
\title{sdc_loglinear_iter}
\usage{
sdc_loglinear_iter(
  data,
  weight,
  varpool,
  numiter = 40,
  epsilon = 0.01,
  fixed_pi = TRUE,
  intermediate_fname = "__loglin_intermediate__.rds",
  restart = FALSE,
  delta = NULL,
  verbose = TRUE,
  blanks_as_missing = TRUE,
  output_filename = NULL
)

\method{print}{sdc_loglinear_iter}(x, summary_outfile = NULL, ...)

\method{plot}{sdc_loglinear_iter}(x, plotpath = NULL, plotvar1 = character(0), plotvar2 = character(0), ...)
}
\arguments{
\item{data}{Data frame containing the data to be evaluated.}

\item{weight}{Column name for sampling weights.}

\item{varpool}{Vector of column names to be used in model.}

\item{numiter}{Maximum number of iterations to run iterative proportional 
fitting for the loglinear model.}

\item{epsilon}{Maximum deviation allowed between observed and fitted margins.}

\item{fixed_pi}{If TRUE, sampling rate assumed to be the same across cells.}

\item{intermediate_fname}{Name of intermediate rds file. At each iteration of
variable selection, the results so far are saved to this file. This file
allows for the process to be restarted if interrupted.}

\item{restart}{If TRUE, restart an interrupted run.}

\item{delta}{Stopping condition for variable selection. If the relative
change in all risk measures is smaller than delta, stop iteration. If NULL
iteration continues till all variables are used or goodness of fit measures
are all negative.}

\item{verbose}{If TRUE, print updates to console at each iteration of the
variable selection process.}

\item{blanks_as_missing}{If TRUE, character and factor variables that are
blank or pure whitespace are treated as missing values.}

\item{output_filename}{Name of the csv file to save the data set with 
record-level risk measures, .tau1_rec and .tau2_rec, attached. NULL if no 
output file is to be saved.}

\item{x}{Object of class sdc_loglinear_iter, as returned by sdc_loglinear.}

\item{summary_outfile}{Name of summary output .txt file. If not NULL, console
output is copied to the file. Default is NULL (no logging of output).
Errors and warnings are not diverted (consider running in batch mode if
logging is needed).}

\item{...}{Currently unused. For NextMethod compatibility.}

\item{plotpath}{Directory to save plots. Plots are saved as \emph{jpeg} files 
(quality = 100\%). If the directory does not exist, it is first created.
If \code{plotpath} is NULL (default), plots are not saved.}

\item{plotvar1}{A vector of names of discrete variables for boxplots. If 
none, boxplots are not produced.}

\item{plotvar2}{A vector of names of continuous variables for scatterplots. 
If none, scatterplots are not produced.}
}
\value{
An object of type \code{sdc_loglinear_iter} containing calculated 
risk measures.
}
\description{
Calculates file-level risk measures using a loglinear model with forward
stepwise variable selection for interaction terms.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: S3 print method for sdc_loglinear_iter objects

Prints summary of iterative loglinear fit.

\item \code{plot}: S3 plot method for \code{sdc_loglinear_iter} 
  objects

Produces boxplots and scatterplots of record-level risk measures, tau1 and
tau2.
}}

\examples{
data(exampledata)
vars <- c("BORNUSA", "CENREG", "DAGE3", "DRACE3", "EDUC3", "GENDER")
wgt <- "WEIGHT"

results <- sdc_loglinear_iter(exampledata, wgt, vars)
print(results)
}
